+++ tests/check/valid/class.rs
@@ -1,11 +1,8 @@
-use mamba::check::check_all;
-use mamba::lex::tokenize;
-use mamba::parse::parse;
-
+use crate::check::{check_test, CheckTestRet};
 use crate::common::resource_content;
 
 #[test]
-fn top_level_tuple() {
+fn top_level_tuple() -> CheckTestRet {
     let source = resource_content(true, &["class"], "top_level_tuple.mamba");
-    check_all(&[(*parse(&tokenize(&source).unwrap()).unwrap(), None, None)]).unwrap();
+    check_test(&source)
 }
