--- tests/check/valid/control_flow.rs
@@ -1,1 +1,0 @@
-
