+++ tests/check/valid/mod.rs
@@ -1,5 +1,3 @@
 pub mod class;
-pub mod control_flow;
 pub mod definition;
 pub mod function;
-pub mod python_primitives;
