+++ tests/output/mod.rs
@@ -3,12 +3,12 @@ use std::process::Command;
 
 use python_parser::ast::Statement;
 
+use mamba::common::delimit::newline_delimited;
 use mamba::pipeline::transpile_directory;
 
 use crate::common::{delete_dir, python_src_to_stmts, resource_content, resource_content_path,
                     resource_content_randomize, resource_path};
 use crate::output::common::PYTHON;
-use mamba::common::delimit::newline_delimited;
 
 mod common;
 
@@ -24,11 +24,11 @@ fn test_directory(
         resource_content_randomize(true, input, &format!("{}.py", file_name));
 
     let res = fallable(valid, input, &output_path, &output_file, file_name);
-    let (check_ast, out_ast) = res?;
     delete_dir(&output_path).map_err(|_| vec![])?;
+    let (check_ast, out_ast) = res?;
 
     // Convert to newline delimited string for more readable diff
-    let check_string= newline_delimited(check_ast.iter().map(|stmt| format!("{:?}", stmt)));
+    let check_string = newline_delimited(check_ast.iter().map(|stmt| format!("{:?}", stmt)));
     let out_string = newline_delimited(out_ast.iter().map(|stmt| format!("{:?}", stmt)));
     assert_eq!(check_string, out_string);
     Ok(())
